# docs/assets/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "TVTime Scrapy API",
        "description": "Unofficial TVTime API via Scrapy",
        "version": "0.1.0"
    },
    "paths": {
        "/scraper/": {
            "post": {
                "tags": [
                    "scraper"
                ],
                "summary": "Start Scrape Task",
                "operationId": "start_scrape_scraper__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TVTimeUser"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "204": {
                        "description": "Scrape Task Accepted"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/scraper/{task_id}": {
            "get": {
                "tags": [
                    "scraper"
                ],
                "summary": "Scrape Task Status",
                "description": "_summary_\n\nArgs:\n    task_id (str): Celery Task ID\n    tvtime_scraper_service (Annotated[TVTimeScraperService, Depends): TVTimeScraperService\n\nReturns:\n    Response: status of the celery task",
                "operationId": "scrape_status_scraper__task_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Task Id",
                            "type": "string"
                        },
                        "name": "task_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "204": {
                        "description": "Scrape Task Accepted"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "TVTimeUser": {
                "title": "TVTimeUser",
                "required": [
                    "username",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "title": "Username",
                        "type": "string"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string"
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    }
}